[S3] Fix error if fixed_bundle is empty

+++ external-import/s3/src/s3.py
@@ -219,6 +219,7 @@ class S3Connector:
         ignored_entities = []
         data = json.loads(bundle)
         new_bundle_objects = []
+        new_bundle = []
         for obj in data["objects"]:
             included_entities.append(obj["id"])
         for obj in data["objects"]:
@@ -336,7 +337,9 @@ class S3Connector:
                 )
                 continue
             new_bundle_objects.append(obj)
-        new_bundle = self.helper.stix2_create_bundle(new_bundle_objects)
+
+        if new_bundle_objects:
+            new_bundle = self.helper.stix2_create_bundle(new_bundle_objects)
         return new_bundle
 
     def process(self):
@@ -354,11 +357,12 @@ class S3Connector:
                 content = data["Body"].read()
                 self.helper.log_info("Sending file " + o.get("Key"))
                 fixed_bundle = self.fix_bundle(content)
-                self.helper.send_stix2_bundle(bundle=fixed_bundle, work_id=work_id)
-                self.helper.log_info("Deleting file " + o.get("Key"))
-                self.s3_client.delete_object(
-                    Bucket=self.s3_bucket_name, Key=o.get("Key")
-                )
+                if fixed_bundle:
+                    self.helper.send_stix2_bundle(bundle=fixed_bundle, work_id=work_id)
+                    self.helper.log_info("Deleting file " + o.get("Key"))
+                    self.s3_client.delete_object(
+                        Bucket=self.s3_bucket_name, Key=o.get("Key")
+                    )
 
             message = (
                 "Connector successfully run ("
